updated merge

+++ database/migrations/2026_01_13_054741_create_supplier_ledgers_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('supplier_ledgers', function (Blueprint $table) {
-            $table->engine = 'InnoDB';
-            $table->id();
-            $table->foreignId('supplier_id')->constrained('suppliers')->onDelete('cascade');
-            $table->nullableMorphs('transactionable'); 
-            $table->date('date');
-            $table->string('description')->nullable();
-            $table->decimal('debit', 15, 2)->default(0); 
-            $table->decimal('credit', 15, 2)->default(0); 
-            $table->decimal('balance', 15, 2)->default(0);
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('supplier_ledgers')) {
+            Schema::create('supplier_ledgers', function (Blueprint $table) {
+                $table->engine = 'InnoDB';
+                $table->id();
+                $table->foreignId('supplier_id')->constrained('suppliers')->onDelete('cascade');
+                $table->nullableMorphs('transactionable'); 
+                $table->date('date');
+                $table->string('description')->nullable();
+                $table->decimal('debit', 15, 2)->default(0); 
+                $table->decimal('credit', 15, 2)->default(0); 
+                $table->decimal('balance', 15, 2)->default(0);
+                $table->unsignedBigInteger('created_by')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
